+++ src/UI/DropDown.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef, useState } from "react";
 import { Link } from "react-router-dom";
 import styles from "./DropDown.module.scss";
-import { ListItem, DropDownProps } from "../interface/Interface";
+import { ListItem } from "../interface/Interface";
 
 const list: ListItem[] = [
   { id: 1, title: "마이페이지" },
@@ -10,27 +10,10 @@ const list: ListItem[] = [
   { id: 4, title: "로그인" },
 ];
 
-function DropDown({ dropDown, setDropDown }: DropDownProps) {
-  const dropdownRef = useRef<HTMLUListElement | null>(null);
-
-  const handleClickOutside = (event: MouseEvent) => {
-    if (
-      dropdownRef.current &&
-      !dropdownRef.current.contains(event.target as Node)
-    ) {
-      setDropDown(false);
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
+function DropDown() {
   return (
     <>
-      <ul ref={dropdownRef} className={styles.drop_list}>
+      <ul className={styles.drop_list}>
         {list.map((item) => (
           <li className={styles.drop_listItem} key={item.id}>
             <Link to={"/"}>
